fixes for windows

+++ libraries/glvisualization/packages/context/src/windowutilities.cpp
@@ -1,6 +1,7 @@
 #include "../inc/windowutilities.hpp"
 
 #include <GLFW/glfw3.h>
+#include <utility>
 
 namespace cie {
 namespace gl {
@@ -9,14 +10,14 @@ namespace gl {
 std::pair<size_t, size_t> getMonitorResolution( GLFWmonitor* monitor )
 {
     const GLFWvidmode* mode = glfwGetVideoMode( monitor );
-    return std::make_pair<size_t, size_t>( mode->width, mode->height );
+    return std::make_pair<size_t, size_t>(size_t(mode->width), size_t(mode->height));
 }
 
 
 std::pair<size_t, size_t> getPrimaryMonitorResolution()
 {
     const GLFWvidmode* mode = glfwGetVideoMode( glfwGetPrimaryMonitor() );
-    return std::make_pair<size_t, size_t>( mode->width, mode->height );
+    return std::make_pair<size_t, size_t>( size_t(mode->width), size_t(mode->height) );
 }
